column name error handling changes
Moved all error handling for column names to one function: check_column_names() which checks if the relevant column names are in their given dataframe. If they are not, it tries to find a potential alias, and if there are none, throws an error. (if the column name is in the dataframe nothing happens)