Updated dependency compatibilities to allow for the most recent versions of dependencies (#142)
* Changed compat to allow for most recent versions

I am doing this to see if the package compiles on the most recent versions of all our dependencies

* Updated compatibility for Julia 1.11

Updated compat statements to allow for the most recent dependency versions which work with Julia 1.11

* Remove compat limitation for linear algebra

Linear algebra had previously been allowed to be any version. I am reverting a previous change on this branch to make that the case again

* Changed Lux compatatability to not include 1.4.2

Lux 1.4.1 allows the package to precompile, but conflicts between NLSolve and Lux 1.4.2 appear to be causing the package to fail to precompile

* Updating Compatability for NonLinearSolveBase

precompilation issue appears to actually be with NonlinearSolveBase versions 1.4 or higher. I am limiting compatability to NonlinearSolveBase versions of 1.3

* fixed incorrect compatibility statement

* Add NonlinearSolveBase as a weak dependency